Initial commit of issue 2066 for OSX
